nnfs/requirements.txt
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
idna==3.8
kaggle==1.6.17
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.32.3
six==1.16.0
text-unidecode==1.3
tqdm==4.66.5
urllib3==2.2.2
webencodings==0.5.1
